backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
pyjwt==2.8.0
cryptography==41.0.7
httpx==0.25.2
python-multipart==0.0.6
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
requests==2.31.0
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
